+++ ROADMAP.zh.md
@@ -3,7 +3,7 @@
 
 
 ## 已完成
-- 清理了 `api_server.py` 和测试中的无用导入。
+- 清理了 `mytimer/server/api.py` 和测试中的无用导入。
 - 安装了所需依赖。
 - 通过 `pytest -q` 验证所有单元测试均通过。
 - 实现了计时器结束通知模块 `Notifier`。
