--- demo_output.txt
@@ -1,17 +1,0 @@
-Starting server...
-Server started.
-
-$ /root/.pyenv/versions/3.12.10/bin/python client_controller.py create 5
-1
-$ /root/.pyenv/versions/3.12.10/bin/python client_controller.py list
-{"1": {"duration": 5.0, "remaining": 5.0, "running": true, "finished": false}}
-$ /root/.pyenv/versions/3.12.10/bin/python client_controller.py tick 3
-ticked
-$ /root/.pyenv/versions/3.12.10/bin/python client_controller.py list
-{"1": {"duration": 5.0, "remaining": 2.0, "running": true, "finished": false}}
-$ /root/.pyenv/versions/3.12.10/bin/python client_controller.py remove 1
-removed
-$ /root/.pyenv/versions/3.12.10/bin/python client_controller.py list
-{}
-
-Stopping server...
