+++ tests/test_api_server.py
@@ -5,7 +5,7 @@ sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from fastapi.testclient import TestClient
 
-from api_server import app, manager, websockets
+from mytimer.server.api import app, manager, websockets
 
 client = TestClient(app)
