+++ tests/test_client_view_layer.py
@@ -9,7 +9,7 @@ import pytest
 def start_server():
     proc = subprocess.Popen([
         "uvicorn",
-        "api_server:app",
+        "mytimer.server.api:app",
         "--host",
         "127.0.0.1",
         "--port",
@@ -35,7 +35,7 @@ def test_client_view_once(start_server):
     timer_id = resp.json()["timer_id"]
 
     result = subprocess.run(
-        [sys.executable, "client_view_layer.py", "--once", "--url", "http://127.0.0.1:8003"],
+        [sys.executable, "-m", "mytimer.client.view_layer", "--once", "--url", "http://127.0.0.1:8003"],
         capture_output=True,
         text=True,
     )
