+++ tests/test_htop_cli.py
@@ -3,6 +3,6 @@ import sys
 
 
 def test_htop_cli_once():
-    result = subprocess.run([sys.executable, 'htop_cli.py', '--once'], capture_output=True, text=True)
+    result = subprocess.run([sys.executable, '-m', 'tools.htop_cli', '--once'], capture_output=True, text=True)
     assert result.returncode == 0
     assert 'Mini htop' in result.stdout
