+++ tools/hello_cli.py
