+++ tools/test_discovery.py
@@ -3,9 +3,9 @@ import subprocess
 import time
 
 async def main():
-    server_proc = subprocess.Popen(['python', 'mock_server.py'])
+    server_proc = subprocess.Popen(['python', '-m', 'tools.mock_server'])
     time.sleep(0.5)  # give the server time to start
-    from server_discovery import discover_server
+    from tools.server_discovery import discover_server
     servers = await discover_server()
     server_proc.terminate()
     print('Servers found:', servers)
